model loading hype

+++ include/vkMaze/Buffers.hpp
@@ -36,6 +36,6 @@ private:
   VulkanContext *ctx;
   VulkanEngine *eng;
   std::vector<Vertex> *vertices;
-  std::vector<uint16_t> *indices;
+  std::vector<uint32_t> *indices;
   FrameData *frame;
 };
